+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 ## Features:
 - Base config class with more control over parameters
+- Support for custom polynomial CPU power models via the `cpuPowerModel` configuration parameter
 
 # Version 1.0.0
 ## Bug Fixes: